+++ CHANGES.rst
@@ -10,6 +10,7 @@ OpenREM version history
 * `#740`_  Installation: fixed Celery version to avoid dependency on Django 1.11
 * `#739`_  Imports: fixed import errors for GE surgical fluoroscopy
 * `#738`_  Logging: added single_date query date to log, added tasks aborts to logs
+* `#737`_  Interface and exports: specify number of events and export to PHE 2019 CT survey specification
 * `#736`_  Query-retrieve: duplicate study level responses now removed from query
 * `#735`_  Imports: switched to more secure defusedxml for parsing XML in comments
 * `#734`_  Query-retrieve: handle illegal image level response with no instance number